[FTR] Challenge 12 - Word Cloud Implementation
Feature Description

Word cloud map generation implementation.  Really the logic is the difficult part
but would come down to what type of words should be considered for generation of
the keys.  Handling of proper nouns would likely require addition argument for the
key generation (i.e. indication if it is a start of a new sentence).

Related Issues
- #25