[FTR] Challenge 21 - Matching Parenthesis Index Implementation
Feature Description

The current implementation just traverses the string from the provided
index until the end of the string or until the total count of open
parenthesis results in a value of 1.  Went ahead and started the count
at one instead of zero to allow to the decrement to occur prior to the
conditional check for matching index.

Related Issues
- #43